requirements.txt
psycopg2-binary==2.9.1
numpy==1.24.3
pandas==2.1.3
shapely==2.0.2